+++ .gitignore
@@ -6,3 +6,7 @@ dist
 *.bak
 .envrc
 .direnv/
+.vscode/
+package-lock.json
+
+
