--- .vscode/bookmarks.json
@@ -1,14 +1,0 @@
-{
-	"files": [
-		{
-			"path": "src/pipe.ts",
-			"bookmarks": [
-				{
-					"line": 165,
-					"column": 1,
-					"label": ""
-				}
-			]
-		}
-	]
-}
